# api-doc.json
{
  "swagger": "2.0",
  "info": {
    "title": "Survey Node API",
    "version": "1.0.0",
    "description": ""
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/api/survey_responses/": {
      "get": {
        "tags": [
          "Responses"
        ],
        "summary": "Get All Responses",
        "description": "",
        "operationId": "getResponses",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Responses"
          }
        }
      }
    },
    "/api/survey_responses/create": {
      "post": {
        "tags": [
          "Responses"
        ],
        "summary": "Create Responses",
        "description": "",
        "operationId": "createResponses",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "responses": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "survey_id": {
                        "type": "number"
                      },
                      "option_id": {
                        "type": "number"
                      },
                      "question_id": {
                        "type": "number"
                      }
                    }
                  }
                }
              },
              "required": [
                "responses"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Responses"
          }
        }
      }
    },
    "/api/survey_responses/{id}": {
      "get": {
        "tags": [
          "Responses"
        ],
        "summary": "Get Response",
        "description": "",
        "operationId": "getResponse",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "number"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Responses"
          }
        }
      },
      "put": {
        "tags": [
          "Responses"
        ],
        "summary": "Update Response",
        "description": "",
        "operationId": "updateResponse",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "response": {
                  "type": "string"
                }
              },
              "required": [
                "response"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Responses"
          }
        }
      },
      "delete": {
        "tags": [
          "Responses"
        ],
        "summary": "Delete Response",
        "description": "",
        "operationId": "deleteResponse",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "number"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Responses"
          }
        }
      }
    },
    "/api/surveys/": {
      "get": {
        "tags": [
          "Surveys"
        ],
        "summary": "Get All Surveys",
        "description": "",
        "operationId": "getSurveys",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Surveys"
          }
        }
      },
      "post": {
        "tags": [
          "Surveys"
        ],
        "summary": "Multipurpose endpoint to create or update a survey, its corresponding questions, and options in one operation",
        "description": "",
        "operationId": "createOrUpdateSurveys",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "survey_name": {
                  "type": "string"
                },
                "survey_description": {
                  "type": "string"
                },
                "is_active": {
                  "type": "boolean"
                },
                "questions": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer"
                      },
                      "question": {
                        "type": "string"
                      },
                      "input_type_id": {
                        "type": "integer"
                      },
                      "admin_id": {
                        "type": "integer"
                      },
                      "options": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "integer"
                            },
                            "label": {
                              "type": "string"
                            },
                            "question_id": {
                              "type": "integer"
                            },
                            "admin_id": {
                              "type": "integer"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              },
              "required": [
                "survey_name",
                "survey_description",
                "is_active"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Surveys"
          }
        }
      }
    },
    "/api/surveys/create": {
      "post": {
        "tags": [
          "Surveys"
        ],
        "summary": "Multipurpose endpoint to create or update a survey, its corresponding questions, and options in one operation",
        "description": "",
        "operationId": "createOrUpdateSurveys",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "survey_name": {
                  "type": "string"
                },
                "survey_description": {
                  "type": "string"
                },
                "is_active": {
                  "type": "boolean"
                },
                "questions": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer"
                      },
                      "question": {
                        "type": "string"
                      },
                      "input_type_id": {
                        "type": "integer"
                      },
                      "admin_id": {
                        "type": "integer"
                      },
                      "options": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "integer"
                            },
                            "label": {
                              "type": "string"
                            },
                            "question_id": {
                              "type": "integer"
                            },
                            "admin_id": {
                              "type": "integer"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              },
              "required": [
                "survey_name",
                "survey_description"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Surveys"
          }
        }
      }
    },
    "/api/surveys/admin/": {
      "get": {
        "tags": [
          "Surveys"
        ],
        "summary": "Get All Surveys created by Admin",
        "description": "",
        "operationId": "getAdminSurveys",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Surveys"
          }
        }
      }
    },
    "/api/surveys/admin/{id}": {
      "get": {
        "tags": [
          "Surveys"
        ],
        "summary": "Get Single Survey create by Admin",
        "description": "",
        "operationId": "getSingleAdminSurvey",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Surveys"
          }
        }
      }
    },
...
    "/api/surveys/{id}/responses": {
      "get": {
        "tags": [
          "Surveys"
        ],
        "summary": "Get Survey Responses",
        "description": "",
        "operationId": "getSurveyResponses",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Surveys"
          }
        }
      }
    },
    "/api/surveys/{id}/{id}": {
      "get": {
        "tags": [
          "Surveys"
        ],
        "summary": "Get Survey",
        "description": "",
        "operationId": "getSurvey",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "Surveys"
          }
        }
      },
      "put": {
        "tags": [
          "Surveys"
        ],
        "summary": "Update specific survey",
        "description": "",
        "operationId": "updateSurveys",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "number"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "survey_name": {
                  "type": "string"
                },
                "survey_description": {
                  "type": "string"
                },
                "is_active": {
                  "type": "boolean"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Surveys"
          }
        }
      },
      "delete": {
        "tags": [
          "Surveys"
        ],
        "summary": "Delete Survey",
        "description": "",
        "operationId": "deleteSurvey",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "number"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Survey deleted"
          }
        }
      }
    }
  }
}